refactored pictures, grouserun, fixed fly path

+++ periphery/gsensor.h
@@ -11,7 +11,7 @@
 #include "adc14_msp432.h"
 #include "gpio_msp432.h"
 
-extern const uint8_t SENSOR_TRIGGER_G;
+extern const uint16_t SENSOR_TRIGGER_G;
 
 class GSensor {
